autobuild-run: use docopt for argument parsing
Using docopt, argument parsing becomes trivial. Adding a new argument is
a matter of updating the usage text.

This commit removes the original cumbersome argument handling and uses
docopt instead. A method is added to import the settings from the
configuration file in a similar dictionary as the one created by
docopt, so that both can be merged (giving priority to the arguments,
as before).

With these changes, each option can be passed as argument and in the
configuration file. This means that http-user and http-password can now
also be added as argument (even though passing the password on the
command-line is not recommended).

[Thomas P.: fix the code to actually give priority to the arguments
and not the configuration file.]

Signed-off-by: Thomas De Schampheleire <thomas.de.schampheleire@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>