Enable paranoid path checking
Now that Buildroot has support for paranoid path checking (check for
clearly invalid header and library directories used when
cross-compiling), let's enable this in the autobuilders.

Without this, the paranoid path checking only emits warnings. With
this patch enabled, whenever an invalid header or library path is
used, the build will stop and fail, which will allow us to clearly
notice such cases.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
---
Do we want to apply this right now? We already have a lot of failures
caused by the static/shared change. Do we want to have more?