autobuild-run: correctly handle default argument values
In the original code sent on the mailing list, priority for arguments was
given to the config file, followed by arguments passed on the command-line.
In this order, the defaults specified to docopt would indeed be treated as
lowest priority, as expected.

However, the arguments specified on the command-line should have priority
over those from the config file, so the expected priority is:
    command-line > config file > defaults

By switching the order of argument dictionaries when calling the merge
method, the defaults are no longer handled correctly. To fix this, we must
not tell docopt about the defaults, but keep them in a separate dictionary.
Then, using a two-step merge we can obtain the right priority for arguments.

For some reason, docopt does not like us passing __doc__ directly if it is
constructed dynamically with %s specifiers. Work around this via an
intermediate variable doc.

Signed-off-by: Thomas De Schampheleire <thomas.de.schampheleire@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>