autbuild-run: fix legal-info
Currently, running legal-info is broken, as it only runs a plain 'make',
which does not do much...

Fix that by explicitly pass 'legal-info' as the make target. Also, we
need to specify the download dir, to avoid using the default one (which
does not exist) or a user-defined one (in the environment).

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>