br-reproduce-build: fix URL of gitid
The canonical URL to the results on the autobuilder server wants the
sha1 to be prefixed with the first three letters of the sha1, like so:

    b52/b52c54eafc2ea8814bb49850823e55d8e034d33d

Currently, this is done by a redirect on the server, but it is broken,
so that requests that do not conform to that scheme return empty
results. See for example the difference between those two calls:

    wget -O OK 'http://autobuild.buildroot.org/results/b2c/b2cefd7d362c12ed99b708d11b988704778d450c/defconfig'
    wget -O KO 'http://autobuild.buildroot.org/results/b2cefd7d362c12ed99b708d11b988704778d450c/defconfig'

So, currently, passing only the sah1 on the command line of br-reproduce
would not work, because the sha1 is used as-is to construct the URL. To
make it work, a user had to pass the sha1 prefixed with the first three
chars.

However, this is not very convenient.

Fix that:
  - internally reconstruct the canonical URL;
  - still accept a prefix-sha1 on the command line (for users that got
    used to that behaviour).

Reported-by: Fabio Porcedda <fabio.porcedda@gmail.com>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>