toolchains/configs: temporarily remove the musl toolchain config
The musl support is still experimental, and there are lots of packages
that don't build with it. So, during the rc/stabilization phase, we
remove this toolchain from the autobuilders tests, in order to focus
on the more important issues.

The musl configuration will be re-enabled once 2015.05 is released,
and hopefully kept continuously during the next cycle.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>