web/import: adjust reason detection logic
Due to the umask handling, the main Buildroot "make" may recurse into
another "make" doing the really work. This wasn't properly handled in
the hacky logic that tries to detect the reason of the build
failure. Therefore, this patch updates this detection code to try to
make it a bit smarter (and uglier).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>