autobuild: add config option to set niceness
When the machine running the autobuilder is shared, it is important
that the autobuilder instances do not hog all of the CPU.

Add an option to the configuration, so that the user can set the
niceness to run the instances with. By default, the niceness is 0,
which means the current niceness.

Properly doing so would require setting up a container with proper
CPU cgroup, but it's a bit complicated.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>