autobuild: add a function to check if the toolchain is usable
Currently, the fixup_config() function validates that the
configuration is valid, and if not either adjusts it or rejects
it. However, if the reason for rejecting the configuration is because
the toolchain is not usable at all on the current machine, we simply
enter an endless loop as the randomization of the package selection
will not make the configuration acceptable.

This commit introduces a new function, is_toolchain_usable() which is
responsible for checking that the toolchain is usable.

That function currently does nothing and always considers the
toolchain to be usable; actual tests will be added in followup
commits.

[Thomas:
 - rename function to 'is_toolchain_usable' because the matter is not
   really about the configuration being fixable.
 - do not add an error message to the logs when the configuration is
   not fixable, since it is not an error.
 - fix typos in the commit log.]

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>