etc: add SysV-init startup script and sample config file
This startup script (and its companion configuration file in
/etc/default/) allows one to easily start/stop an autobuilder.

An autobuilder can be run either on the current system, or in
a chroot (preferred). Running in a chroot is done by setting
the AUTOBUILD_CHROOT variable in the configuration file. If
that variable is not set, then the autobuilder does not enter
a chroot.

It is expected that the buildroot-test tree be already present
somewhere (for example, as a git checkout).

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>