br-reproduce-build: do olddefconfig before starting the build
Since the host configuration can be different between the autobuilder
slave and the machine where we try to reproduce the build failure, the
set of available options might be different. Due to this, a .config
that is "valid" for a given machin may have missing options for
another machine.

A good example is the BR2_GOOGLE_BREAKPAD_ENABLE option. Since it
depends on BR2_HOST_GCC_AT_LEAST_4_8, this option will *not* be
present in a .config file generated on a machine that has a host gcc
older than 4.8. However, when running br-reproduce-build on a machine
with a gcc newer than 4.8, the option is available, but has no value
defined in the .config.

To cope with this, we simply run an "olddefconfig". This means we
aren't rebuilding *exactly* the same configuration, but it's the best
approximation we can do.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>