toolchains/configs: make external toolchain explicit
Some of the toolchain configs rely on the default to select which
external toolchain to use. However, this is wrong for two reasons:
- when the defaults change in Buildroot, the toolchain config will
  change under the hood;
- when the autobuild-run script adds some options (in particular,
  BR2_STATIC_LIBS), it is possible that the default changes (or is
  no longer available).

Both can be fixed by explicitly adding the external toolchain option
we want to the config file. Indeed, the autobuild-run script will
then detect that there is a difference between the generated config
and the base one, and will discard it.

Fixes:
http://autobuild.buildroot.net/results/39888c188c0d13219a8419897a833275fcc81597

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Cc: Thomas De Schampheleire <thomas.de_schampheleire@nokia.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>