web/import.inc.php: do not make the build-time.log file mandatory
Some failures happen early in the build process (such as the "XYZ is
in the dependency chain of ABS" error), even before a build-time.log
is generated. Currently, such build results are discarded, which isn't
nice.

So, let's accept results without a build-time.log, and assume their
duration was 0 seconds.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>