autobuild-run: use in-tree genrandconfig script (initial version)
The Buildroot tree now has a genrandconfig script that replaces a
large part of the autobuild-run script. Call this script rather than
having it in autobuild-run.

The genrandconfig script inside the Buildroot tree has the advantage
that it can be adapted to the changes in the source, without the
need to update the autobuild-run script all the time. Also, it makes
it feasible to support autobuilding multiple branches.

Signed-off-by: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>