autobuild-run: adapt to genrandconfig with toolchains-csv option
The --toolchains-url option has been removed from genrandconfig, it has
been replaced with --toolchains-csv.

We don't actually need to pass this option if not set, so that the
in-tree default will be used. This makes sure that the autobuild-run
doesn't need to be updated if we change the location of the CSV file.

The -i option has been removed from genrandconfig, it has been replaced
with an explicit -o OUTPUTDIR and -b BUILDROOTDIR.

Signed-off-by: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>